base for metrics collector test
